swan command skeleton
This patch adds the (currently empty) `swan` command-line utility.  This
will be the main way to interact with Swanson's compiler and runtime.
It will be structured as a single executable with many “subcommands”,
just like `git` and `svn`.  Right now the `swan` command is empty; there
aren't any subcommands defined.