Evaluation of S0 files
This patch adds a new `swan s0 evaluate` command, which reads in an S0
file and evaluates it.  Right now, we only support evaluating in the
context of the Swanson “kernel”, which defines a minimal set of objects
in C.  The hope is that the bulk of the Swanson prelude will be written
in Swanson itself.