Starting the kernel
This patch starts the Swanson kernel.  The “prelude” is going to be the
full set of built-in types and values that you get for free.  (The core
of the standard library, if you will.)  My intention is that (most of)
the prelude will be written in pure Swanson.  The “kernel” is going to
be the subset that you have to write in the bootstrap language (in our
case, C) to provide everything needed to define the prelude.

The bulk of this patch is a set of helper methods for defining the
kernel objects in C, along with documentation for those objects.