Explicit heap allocation of values
This patch adds support for allocating values directly from the heap.
These values are immediately freed when you call their `~unref` method.
(You cannot create aliases to these values, since they don't have an
`~alias` method.  That means we can guarantee that there's only ever one
reference to the value, allowing us to immediately free it when that
reference is dropped.)

To test this, I've added a `+` method to the kernel's `size` type, which
allocates its result from the heap.

This patch also reorganizes the helper macros that we've been using to
write the kernel.  They're no longer available in a public header file.
The intention is that they're only needed for the kernel; third-party C
libraries will be integrated into Swanson using a different FFI-like
mechanism.