Simple strings
This patch adds a new type for storing string data.  The content of a
string value is a `swan_string` instance, which contains a pointer and a
size; this will let us support slices without having to copy data, since
the pointer and size can refer to any subset of a buffer.