S0 parser
This patch adds a parser for the S0 language.  The S0 parser is
implemented in pure C.  You provide a set of callback functions that
will be called as the statements in the S0 file are encountered and
parsed.

This patch also adds a "swan s0 verify" command, which does a simple
syntax-check of an S0 file.  We print out a JSON representation of each
statement in the file on a separate line.  This serves as the basis for
a pretty exhaustive list of test cases for this initial draft of the S0
language.