Require all values to be unreferenced
You're now obligated to unreference any of the values that you create or
alias in a block of S₀ code.  This doesn't necessarily mean that you
have to call the `~unref` operation directly; one of the other
operations you call can unreference the value for you.  The specific
requirement is that every named value must be empty when the chunk of
code terminates.