Size type and heap region
This patch starts our support for “types” and “regions”.  In Swanson, a
type does not concern itself with memory management; the type is
represented by the set of operations that can be performed on instances
of that type, regardless of how the instances were allocated.

A “region” knows how to allocate objects in a certain way.  Currently we
only have a region that allocates objects from the heap.  The region has
an operation that takes in a type, and returns an “allocator” for that
type.  The allocator can be used to create new instances of the type,
that live within the corresponding region.  The region provides the
memory management operations for the new instances, while the type
provides all of the other operations.