Fixed-size arrays
This patch adds a new `kernel.array` function, which creates a new type
to hold a fixed-size array of some other type.  For memory management
purposes, currently you can only have a single reference to the array or
to any of the arrays elements.  The array's `claim` operation allows you
to “trade in” your reference to the array for a reference to one of its
elements.  And each element's `return` operation lets you trade in your
reference to an element for a reference to the original array.  (We need
this to allow you to work with arrays that can't be aliased.)