Value representations
We now keep track of a “representation” for each Swanson value.  The
representation is an indication of what C type the value content pointer
refers to.  This lets us implement more rigorous checks within our C
operation implementations, without having to place restrictions of the
particular operation sets for the values.  (That's how we implemented
the checks previously; a value was considered a size iff it used our
predefined `swan_size` operation set.)