Array claim operation (again)
This patch turns an array's `[]` operation into a “claim”, rather than a
“get”.  This replaces your reference to the array with a reference to
one of the array's elements.  You must relinquish your reference to the
element to reclaim your reference to the array.  This ensures that, at
any given time, you can only have one active reference to any portion of
the array.