“self” parameters appear first
We used to assume that any “output” parameters for an operation would
appear first in the parameter list — even before any “self” parameter
for the value that the operation belongs to.  This patch reorders things
so that the self parameter comes first.  This lets us use the

    output = block:"()"(input1, input2);

syntactic sugar when calling blocks.  (The block's `()` operation needs
the block to appear first, since it passes any remaining parameters
unchanged to the block itself.)